Isolate bank bundle simulation and tests
